+++ Trainer_v5/Trainer.Source/DetailWindowTrainer.cs
@@ -1,6 +1,6 @@
 ﻿using System.Collections.Generic;
+using Trainer_v5.Trainer.Source.Window;
 using Trainer_v5.Window;
-using UnityEngine.UI;
 using UnityEngine;
 
 namespace Trainer_v5
@@ -22,6 +22,7 @@ namespace Trainer_v5
 				UIFactory.Button("Demand", () => EmployeeDemandChangeWindow.Instance.Show()),
 				// UIFactory.Button("Creativity",  SetCreativity),
 				UIFactory.Button("Inspiration", SetInspiration),
+				UIFactory.Button("LeadSpec", () => EmployeeLeadSpecChangeWindow.Instance.Show()),
 			};
 
 			// add components to DetailWindow