+++ Trainer_v5/Trainer.Source/UIFactory.cs
@@ -59,6 +59,14 @@ namespace Trainer_v5
 			toggle.onValueChanged.AddListener(action);
 			return toggle;
 		}
+
+		public static InputField
+		MultilineInput()
+		{
+			var r = WindowManager.SpawnInputbox();
+			r.lineType = InputField.LineType.MultiLineNewline;
+			return r;
+		}
 	}
 
 	public class TextStyle